# Vectorize `bitset` from string construction