--- a/stl/inc/bitset
+++ b/stl/inc/bitset
@@ -120,8 +120,7 @@ private:
     template <class _Traits, class _Elem>
     _CONSTEXPR23 void _Construct(const _Elem* const _Ptr, size_t _Count, const _Elem _Elem0, const _Elem _Elem1) {
 #if _USE_STD_VECTOR_ALGORITHMS
-        if constexpr (_Bits >= _Bitset_vector_threshold
-                      && _Is_specialization_v<_Traits, char_traits> && sizeof(_Elem) <= 2) {
+        if constexpr (_Is_specialization_v<_Traits, char_traits> && sizeof(_Elem) <= 2) {
             bool _Result;
 
             if constexpr (sizeof(_Elem) == 1) {
@@ -487,6 +486,7 @@ public:
     _CONSTEXPR23 void _To_string(
         _Elem* const _Buf, const size_t _Len, const _Elem _Elem0, const _Elem _Elem1) const noexcept {
 #if _USE_STD_VECTOR_ALGORITHMS
+        constexpr size_t _Bitset_vector_threshold = 32;
         if constexpr (_Bits >= _Bitset_vector_threshold && is_integral_v<_Elem> && sizeof(_Elem) <= 2) {
             if (!_Is_constant_evaluated()) {
                 if constexpr (sizeof(_Elem) == 1) {
@@ -511,9 +511,8 @@ public:
 private:
     friend hash<bitset<_Bits>>;
 
-    static constexpr size_t _Bitset_vector_threshold = 32;
-    static constexpr ptrdiff_t _Bitsperword          = CHAR_BIT * sizeof(_Ty);
-    static constexpr ptrdiff_t _Words = _Bits == 0 ? 0 : (_Bits - 1) / _Bitsperword; // NB: number of words - 1
+    static constexpr ptrdiff_t _Bitsperword = CHAR_BIT * sizeof(_Ty);
+    static constexpr ptrdiff_t _Words       = _Bits == 0 ? 0 : (_Bits - 1) / _Bitsperword; // NB: number of words - 1
 
     _CONSTEXPR23 void _Trim() noexcept { // clear any trailing bits in last word
         constexpr bool _Work_to_do = _Bits == 0 || _Bits % _Bitsperword != 0;