--- a/stl/inc/bitset
+++ b/stl/inc/bitset
@@ -23,6 +23,10 @@ _STL_DISABLE_CLANG_WARNINGS
 #endif // !defined(_STD_BITSET_TO_STREAM_STACK_RESERVATION)
 
 #if _USE_STD_VECTOR_ALGORITHMS
+// These bitset functions sometimes assume bit array has zero padding to multiple of 2 or 4 bytes
+// The assumptions hold true even for vNext suggestion to use smaller types for small bitsets
+// due to vectorization thresholds.
+
 extern "C" {
 __declspec(noalias) void __stdcall __std_bitset_to_string_1(
     char* _Dest, const void* _Src, size_t _Size_bits, char _Elem0, char _Elem1) noexcept;