--- a/stl/inc/bitset
+++ b/stl/inc/bitset
@@ -127,64 +127,64 @@ private:
         constexpr size_t _Bitset_from_string_vector_threshold = 16;
         if constexpr (_Bits >= _Bitset_from_string_vector_threshold
                       && _Is_specialization_v<_Traits, char_traits> && sizeof(_Elem) <= 2) {
-            bool _Result;
+            if (!_STD _Is_constant_evaluated()) {
+                bool _Result;
 
-            if constexpr (sizeof(_Elem) == 1) {
-                _Result = __std_bitset_from_string_1(_Array, reinterpret_cast<const char*>(_Ptr), sizeof(_Array), _Bits,
-                    _Count, static_cast<char>(_Elem0), static_cast<char>(_Elem1));
-            } else {
-                _STL_INTERNAL_STATIC_ASSERT(sizeof(_Elem) == 2);
-                _Result = __std_bitset_from_string_2(_Array, reinterpret_cast<const wchar_t*>(_Ptr), sizeof(_Array),
-                    _Bits, _Count, static_cast<wchar_t>(_Elem0), static_cast<wchar_t>(_Elem1));
-            }
-
-            if (!_Result) {
-                _Xinv();
-            }
-        } else
-#endif // _USE_STD_VECTOR_ALGORITHMS
-        {
-            if (_Count > _Bits) {
-                for (size_t _Idx = _Bits; _Idx < _Count; ++_Idx) {
-                    const auto _Ch = _Ptr[_Idx];
-                    if (!_Traits::eq(_Elem1, _Ch) && !_Traits::eq(_Elem0, _Ch)) {
-                        _Xinv();
-                    }
+                if constexpr (sizeof(_Elem) == 1) {
+                    _Result = __std_bitset_from_string_1(_Array, reinterpret_cast<const char*>(_Ptr), sizeof(_Array),
+                        _Bits, _Count, static_cast<char>(_Elem0), static_cast<char>(_Elem1));
+                } else {
+                    _STL_INTERNAL_STATIC_ASSERT(sizeof(_Elem) == 2);
+                    _Result = __std_bitset_from_string_2(_Array, reinterpret_cast<const wchar_t*>(_Ptr), sizeof(_Array),
+                        _Bits, _Count, static_cast<wchar_t>(_Elem0), static_cast<wchar_t>(_Elem1));
                 }
 
-                _Count = _Bits;
+                if (!_Result) {
+                    _Xinv();
+                }
+            }
+        }
+#endif // _USE_STD_VECTOR_ALGORITHMS
+        if (_Count > _Bits) {
+            for (size_t _Idx = _Bits; _Idx < _Count; ++_Idx) {
+                const auto _Ch = _Ptr[_Idx];
+                if (!_Traits::eq(_Elem1, _Ch) && !_Traits::eq(_Elem0, _Ch)) {
+                    _Xinv();
+                }
             }
 
-            size_t _Wpos = 0;
-            if (_Count != 0) {
-                size_t _Bits_used_in_word = 0;
-                auto _Last                = _Ptr + _Count;
-                _Ty _This_word            = 0;
-                do {
-                    --_Last;
-                    const auto _Ch = *_Last;
-                    _This_word |= static_cast<_Ty>(_Traits::eq(_Elem1, _Ch)) << _Bits_used_in_word;
-                    if (!_Traits::eq(_Elem1, _Ch) && !_Traits::eq(_Elem0, _Ch)) {
-                        _Xinv();
-                    }
+            _Count = _Bits;
+        }
 
-                    if (++_Bits_used_in_word == _Bitsperword) {
-                        _Array[_Wpos] = _This_word;
-                        ++_Wpos;
-                        _This_word         = 0;
-                        _Bits_used_in_word = 0;
-                    }
-                } while (_Ptr != _Last);
+        size_t _Wpos = 0;
+        if (_Count != 0) {
+            size_t _Bits_used_in_word = 0;
+            auto _Last                = _Ptr + _Count;
+            _Ty _This_word            = 0;
+            do {
+                --_Last;
+                const auto _Ch = *_Last;
+                _This_word |= static_cast<_Ty>(_Traits::eq(_Elem1, _Ch)) << _Bits_used_in_word;
+                if (!_Traits::eq(_Elem1, _Ch) && !_Traits::eq(_Elem0, _Ch)) {
+                    _Xinv();
+                }
 
-                if (_Bits_used_in_word != 0) {
+                if (++_Bits_used_in_word == _Bitsperword) {
                     _Array[_Wpos] = _This_word;
                     ++_Wpos;
+                    _This_word         = 0;
+                    _Bits_used_in_word = 0;
                 }
-            }
+            } while (_Ptr != _Last);
 
-            for (; _Wpos <= _Words; ++_Wpos) {
-                _Array[_Wpos] = 0;
+            if (_Bits_used_in_word != 0) {
+                _Array[_Wpos] = _This_word;
+                ++_Wpos;
             }
+        }
+
+        for (; _Wpos <= _Words; ++_Wpos) {
+            _Array[_Wpos] = 0;
         }
     }