--- a/stl/inc/bitset
+++ b/stl/inc/bitset
@@ -142,6 +142,8 @@ private:
                 if (!_Result) {
                     _Xinv();
                 }
+
+                return;
             }
         }
 #endif // _USE_STD_VECTOR_ALGORITHMS