--- a/stl/src/vector_algorithms.cpp
+++ b/stl/src/vector_algorithms.cpp
@@ -3696,12 +3696,108 @@ extern "C" {
 
 __declspec(noalias) bool __stdcall __std_bitset_from_string_1(void* _Dest, const char* _Src, size_t _Size_bytes,
     size_t _Size_bits, size_t _Size_chars, char _Elem0, char _Elem1) noexcept {
-    return __std_bitset_from_string_fallback(_Dest, _Src, _Size_bytes, _Size_bits, _Size_chars, _Elem0, _Elem1);
+    if (_Use_sse42()) {
+        const __m128i _Dx0 = _mm_shuffle_epi8(_mm_cvtsi32_si128(_Elem0), _mm_setzero_si128());
+        const __m128i _Dx1 = _mm_shuffle_epi8(_mm_cvtsi32_si128(_Elem1), _mm_setzero_si128());
+
+        const char* _Src_end = _Src + _Size_chars;
+
+        uint16_t* _Dst_words           = reinterpret_cast<uint16_t*>(_Dest);
+        uint16_t* const _Dst_words_end = _Dst_words + _Size_bytes / sizeof(uint16_t);
+
+        const __m128i _Shuf = _mm_set_epi8(0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15);
+
+        for (;;) {
+            __m128i _Val = _mm_undefined_si128();
+
+            if (const size_t _Left = _Src_end - _Src; _Left > 16) {
+                _Src_end -= 16;
+                _Val = _mm_loadu_si128(reinterpret_cast<const __m128i*>(_Src_end));
+            } else if (_Left == 0) {
+                if (_Dst_words != _Dst_words_end) {
+                    _CSTD memset(_Dst_words, 0, (_Dst_words_end - _Dst_words) * sizeof(uint16_t));
+                }
+                break;
+            } else {
+                _Src_end = _Src;
+                char _Tmp[16];
+                _mm_storeu_si128(reinterpret_cast<__m128i*>(_Tmp), _Dx0);
+                char* const _Tmpd = _Tmp + (16 - _Left);
+                _CSTD memcpy(_Tmpd, _Src_end, _Left);
+                _Val = _mm_loadu_si128(reinterpret_cast<const __m128i*>(_Tmp));
+            }
+
+            const __m128i _Ex1 = _mm_cmpeq_epi8(_Val, _Dx1);
+            const __m128i _Ex0 = _mm_xor_si128(_Val, _Dx0);
+
+            if (!_mm_testc_si128(_Ex1, _Ex0)) {
+                return false;
+            }
+
+            if (_Dst_words != _Dst_words_end) {
+                const __m128i _Ex2 = _mm_shuffle_epi8(_Ex1, _Shuf);
+                *_Dst_words        = static_cast<uint16_t>(_mm_movemask_epi8(_Ex2));
+                ++_Dst_words;
+            }
+        }
+
+        return true;
+    } else {
+        return __std_bitset_from_string_fallback(_Dest, _Src, _Size_bytes, _Size_bits, _Size_chars, _Elem0, _Elem1);
+    }
 }
 
 __declspec(noalias) bool __stdcall __std_bitset_from_string_2(void* _Dest, const wchar_t* _Src, size_t _Size_bytes,
     size_t _Size_bits, size_t _Size_chars, wchar_t _Elem0, wchar_t _Elem1) noexcept {
-    return __std_bitset_from_string_fallback(_Dest, _Src, _Size_bytes, _Size_bits, _Size_chars, _Elem0, _Elem1);
+    if (_Use_sse42()) {
+        const __m128i _Dx0 = _mm_set1_epi16(_Elem0);
+        const __m128i _Dx1 = _mm_set1_epi16(_Elem1);
+
+        const wchar_t* _Src_end = _Src + _Size_chars;
+
+        uint8_t* _Dst_words           = reinterpret_cast<uint8_t*>(_Dest);
+        uint8_t* const _Dst_words_end = _Dst_words + _Size_bytes / sizeof(uint8_t);
+
+        const __m128i _Shuf = _mm_set_epi8(-1, -1, -1, -1, -1, -1, -1, -1, 0, 2, 4, 6, 8, 10, 12, 14);
+
+        for (;;) {
+            __m128i _Val = _mm_undefined_si128();
+
+            if (const size_t _Left = _Src_end - _Src; _Left > 8) {
+                _Src_end -= 8;
+                _Val = _mm_loadu_si128(reinterpret_cast<const __m128i*>(_Src_end));
+            } else if (_Left == 0) {
+                if (_Dst_words != _Dst_words_end) {
+                    _CSTD memset(_Dst_words, 0, (_Dst_words_end - _Dst_words) * sizeof(uint8_t));
+                }
+                break;
+            } else {
+                _Src_end = _Src;
+                wchar_t _Tmp[8];
+                _mm_storeu_si128(reinterpret_cast<__m128i*>(_Tmp), _Dx0);
+                wchar_t* const _Tmpd = _Tmp + (8 - _Left);
+                _CSTD memcpy(_Tmpd, _Src_end, _Left * sizeof(wchar_t));
+                _Val = _mm_loadu_si128(reinterpret_cast<const __m128i*>(_Tmp));
+            }
+
+            const __m128i _Ex1 = _mm_cmpeq_epi16(_Val, _Dx1);
+            const __m128i _Ex0 = _mm_xor_si128(_Val, _Dx0);
+
+            if (!_mm_testc_si128(_Ex1, _Ex0)) {
+                return false;
+            }
+
+            if (_Dst_words != _Dst_words_end) {
+                const __m128i _Ex2 = _mm_shuffle_epi8(_Ex1, _Shuf);
+                *_Dst_words        = static_cast<uint8_t>(_mm_movemask_epi8(_Ex2));
+                ++_Dst_words;
+            }
+        }
+
+        return true;
+    } else {
+        return __std_bitset_from_string_fallback(_Dest, _Src, _Size_bytes, _Size_bits, _Size_chars, _Elem0, _Elem1);
+    }
 }
 
 } // extern "C"