--- a/stl/src/vector_algorithms.cpp
+++ b/stl/src/vector_algorithms.cpp
@@ -3696,6 +3696,7 @@ extern "C" {
 
 __declspec(noalias) bool __stdcall __std_bitset_from_string_1(void* _Dest, const char* _Src, size_t _Size_bytes,
     size_t _Size_bits, size_t _Size_chars, char _Elem0, char _Elem1) noexcept {
+#ifndef _M_ARM64EC
     if (_Use_avx2() && _Size_bits >= 256) {
         _Zeroupper_on_exit _Guard; // TRANSITION, DevCom-10331414
 
@@ -3791,13 +3792,16 @@ __declspec(noalias) bool __stdcall __std_bitset_from_string_1(void* _Dest, const
         }
 
         return true;
-    } else {
+    } else
+#endif // !defined(_M_ARM64EC)
+    {
         return __std_bitset_from_string_fallback(_Dest, _Src, _Size_bytes, _Size_bits, _Size_chars, _Elem0, _Elem1);
     }
 }
 
 __declspec(noalias) bool __stdcall __std_bitset_from_string_2(void* _Dest, const wchar_t* _Src, size_t _Size_bytes,
     size_t _Size_bits, size_t _Size_chars, wchar_t _Elem0, wchar_t _Elem1) noexcept {
+#ifndef _M_ARM64EC
     if (_Use_avx2() && _Size_bits >= 256) {
         _Zeroupper_on_exit _Guard; // TRANSITION, DevCom-10331414
 
@@ -3895,7 +3899,9 @@ __declspec(noalias) bool __stdcall __std_bitset_from_string_2(void* _Dest, const
         }
 
         return true;
-    } else {
+    }
+#endif // !defined(_M_ARM64EC)
+    else {
         return __std_bitset_from_string_fallback(_Dest, _Src, _Size_bytes, _Size_bits, _Size_chars, _Elem0, _Elem1);
     }
 }