--- a/stl/src/vector_algorithms.cpp
+++ b/stl/src/vector_algorithms.cpp
@@ -3671,7 +3671,7 @@ namespace {
             if (_Size_chars > _Size_bits) {
                 _Size_convert = _Size_bits;
 
-                for (size_t _Ix = 0, _Mx = _Size_chars - _Size_bits; _Ix < _Mx; ++_Ix) {
+                for (size_t _Ix = _Size_bits, _Mx = _Size_chars; _Ix < _Mx; ++_Ix) {
                     if (const _Elem _Cur = _Src[_Ix]; _Cur != _Elem0 && _Cur != _Elem1) [[unlikely]] {
                         return false;
                     }
@@ -3681,7 +3681,7 @@ namespace {
             _CSTD memset(_Dest, 0, _Size_bytes);
 
             for (size_t _Ix = 0; _Ix != _Size_convert; ++_Ix) {
-                const _Elem _Cur = _Src[_Size_chars - _Ix - 1];
+                const _Elem _Cur = _Src[_Size_convert - _Ix - 1];
 
                 if (_Cur != _Elem0 && _Cur != _Elem1) [[unlikely]] {
                     return false;