@AlexGuteniev 2024-10-08
> and a single 0.91 scenario is nothing to worry about. 

This is one of the four 15 cases which are just below vectorization threshold. Just one of them has exactly 1.00 as expected